Jugsaw

APIs

Jugsaw.generate_code — Method

generate_code(lang, endpoint::String, appname::Symbol, fcall::JugsawADT, democall::JugsawIR.Call)

Generate code for target language.

Arguments

lang can be a string or an AbstractLang instance that specifies the target language.

Please use subtypes(AbstractLang) for supported client languages.

endpoint is the url for service provider, e.g. it can be https://www.jugsaw.co.

appname is the application name.

fcall is a JugsawADT that specifies the function call.

democall is the demo instance of that function call.

Jugsaw.@register — Macro

@register app expression

Register a function to the application. A function can be registered as a demo, which can take the following forms.

@register app f(args...; kwargs...) == result    # a function call + a test
@register app f(args...; kwargs...) ≈ result     # similar to the above
@register app f(args...; kwargs...)::T           # a function call with assertion of the return type
@register app f(args...; kwargs...)              # a function call
@register app begin ... end                      # a sequence of function

The @register macro checks and executes the expression. If the tests and type asserts in the expression does not hold, an error will be thrown. Otherwise, both the top level function call and those appear in the input arguments will be registered.

Jugsaw.AppSpecification — Type

struct AppSpecification

The application specification.

Fields

name::Symbol

method_names::Vector{String}

method_demos::Dict{String, Vector{JugsawDemo}}

Jugsaw.NoDemoException — Type

struct NoDemoException <: Exception

This error was thrown when a demo matching the target type signature is not found.

Fields

func_sig::Any

methods::Any