Jugsaw.Client

Jugsaw.Client.call — Method

call(
    context::Jugsaw.Client.ClientContext,
    demo::Jugsaw.Client.Demo,
    args...;
    kwargs...
) -> Jugsaw.Client.LazyReturn

Launch a function call.

Jugsaw.Client.healthz — Method

healthz(context::Jugsaw.Client.ClientContext) -> Any

Check the status of the application.

Jugsaw.Client.request_app — Method

request_app(
    context::Jugsaw.Client.ClientContext,
    appname::Symbol
) -> Jugsaw.Client.App

Request an application from an endpoint.

Arguments

context is a ClientContext instance, which contains contextual information like the endpoint.

appname specificies the application to be fetched.

Jugsaw.Client.App — Type

struct App

The Jugsaw application instance.

Note

The Base.getproperty function has been overloaded to favor fetching demos. To get fields, please use app[fieldname]. For example, to get the application name, one should use app[:name].

Fields

name::Symbol

method_demos::OrderedCollections.OrderedDict{Symbol, Vector{Jugsaw.Client.Demo}}

type_table::TypeTable

context::Jugsaw.Client.ClientContext

Jugsaw.Client.ClientContext — Type

mutable struct ClientContext

Fields

endpoint::String

localurl::Bool

project::String

appname::Symbol

version::String

Jugsaw.Client.Demo — Type

struct Demo

Fields

fcall::Call

result::Any

meta::Dict{String}

Jugsaw.Client.DemoRef — Type

struct DemoRef

Fields

demo::Jugsaw.Client.Demo

context::Jugsaw.Client.ClientContext

Jugsaw.Client.DemoRefs — Type

struct DemoRefs

Fields

name::Symbol

demos::Vector{Jugsaw.Client.Demo}

context::Jugsaw.Client.ClientContext

Jugsaw.Client.LazyReturn — Type

struct LazyReturn

A callable lazy result. To fetch the result value, please use lazyresult().

Fields

context::Jugsaw.Client.ClientContext

job_id::String

demo_result::Any