JugsawIR

JugsawIR is an intermediate representation (IR) for exposing and using cloud scientific applications. Its grammar is compatible with JSON at the time of writing, however, it might undergo a refactor to support richer features. The final form should be a real programming language for Web Virtual Machine.

Jugsaw IR can represent data, data types, and function calls. The basic rule is representing a Jugsaw object as a JSON object with extra constaints,

Integers, floating point numbers, Nothing, Missing, UndefInitializer, Symbol and String are directly representable.

Generic objects are represented by a JSON object with at most two fields: fields and type (optional).

Some objects are specialized, including Array, Tuple, Dict, DataType and Enum.

1: Representing Data

Generic Data Types

julia> julia2ir(1.0+2im)[1] |> println
{"fields":[1.0,2.0],"type":"Base.Complex{Core.Float64}"}

Or equivalently, as

{"type":"Base.Complex{Core.Float64}","fields":[1.0,2.0]}

Or when calling a remote function, one can ommit the "type" specification, since the remote already has a copy of data types.

{"fields":[2,3]}

Note

The julia2ir function returns a two element tuple, a representation of object, and a TypeTable to delare types.

Customized Data Types

To customize a Jugsaw data type for your own data type, the following two functions should be overloaded.

native2jugsaw is for converting a native Julia object to a Jugsaw compatible object.

construct_object is for reconstructing the native Julia object from an object of type JugsawADT. The return value must have the same data type as the second argument.

Dictionary

A dictionary is parsed to a JugsawIR.JDict instance, which has two fields keys and vals.

Array

An array is parsed to a JugsawIR.JArray instance, which has two fields size and storage.

Enum

An enum instance is parsed to a JugsawIR.JEnum instance, which has three fields kind, value and options.

2: Representing Data Type

An data type is parsed to a JugsawIR.JDataType instance, which has three fields name, fieldnames and fieldtypes. For example, to represent a complex number type, we can create the following IR

julia> julia2ir(ComplexF64)[1] |> println
{"fields":["Base.Complex{Core.Float64}",["re","im"],["Core.Float64","Core.Float64"]],"type":"JugsawIR.JDataType"}

For convenience, JugsawIR returns a TypeTable instance to represent the types used in parsing.

3: Representing Function Call

A function call is represented as a Jugsaw object with three fields fname, args and kwargs.

julia> fc = Call(sin, (2.0,), (;))
sin(2.0; )

julia> julia2ir(fc)[1] |> println
{"fields":[{"fields":[],"type":"Base.sin"},{"fields":[2.0],"type":"Core.Tuple{Core.Float64}"},{"fields":[],"type":"Core.NamedTuple{(), Core.Tuple{}}"}],"type":"JugsawIR.Call{Base.sin, Core.Tuple{Core.Float64}, Core.NamedTuple{(), Core.Tuple{}}}"}

It is not different with regular Jugsaw object, except that it can executed when it is used to represent a remote call request.

The Grammar

The grammar of JugsawIR in the EBNF format, which can be parsed by lark in Python, Lerche.jl in Julia and hopefully lark-js in Javascript.

object: genericobj1
        | genericobj2
        | genericobj3
        | list
        | string
        | number
        | true
        | false
        | null

genericobj1 : "{" "\"fields\"" ":" list "}"
genericobj2 : "{" "\"type\"" ":" ESCAPED_STRING "," "\"fields\"" ":" list "}"
genericobj3 : "{" "\"fields\"" ":" list "," "\"type\"" ":" ESCAPED_STRING "}"

list : "[" [object ("," object)*] "]"
string : ESCAPED_STRING
number : SIGNED_NUMBER
true : "true"
false : "false"
null : "null"

%import common.ESCAPED_STRING
%import common.SIGNED_NUMBER
%import common.WS
%ignore WS

APIs

JugsawIR.construct_object — Function

construct_object(adt::JugsawADT, demo_object)

Reconstruct the native Julia object from an object of type JugsawADT. The return value must have the same data type as the second argument.

JugsawIR.ir2julia — Method

ir2julia(str::String, demo) -> Any

Convert Jugsaw IR to julia object, given a demo object as a reference. Please check julia2ir for the inverse map.

Examples

julia> JugsawIR.ir2julia("{\"fields\" : [3, 4]}", 1+2im)
3 + 4im

JugsawIR.julia2ir — Method

julia2ir(obj) -> Tuple{Any, Any}

Convert julia object to Jugsaw IR and a type table, where the type table is a special Jugsaw IR that stores the type definitions. Please check ir2julia for the inverse map.

Examples

julia> ir, typetable = JugsawIR.julia2ir(1+2im);

julia> ir
"{\"fields\":[1,2],\"type\":\"Base.Complex{Core.Int64}\"}"

JugsawIR.native2jugsaw — Method

native2jugsaw(object)

Convert a native Julia object to a Jugsaw compatible object.

JugsawIR.Call — Type

struct Call

Fields

fname::Any

args::Tuple

kwargs::NamedTuple

JugsawIR.JArray — Type

struct JArray{T}

The data type for arrays in Jugsaw.

Fields

size::Vector{Int64}

storage::Vector

JugsawIR.JDataType — Type

struct JDataType

The type for specifying data type in Jugsaw.

Fields

name::String

fieldnames::Vector{String}

fieldtypes::Vector{String}

JugsawIR.JDict — Type

struct JDict{K, V}

The dictionary type in Jugsaw.

Fields

keys::Vector

vals::Vector

JugsawIR.JEnum — Type

struct JEnum

The enum type in Jugsaw.

Fields

kind::String

value::String

options::Vector{String}

JugsawIR.JugsawADT — Type

struct JugsawADT

JugsawObject(typename::String, fields::Vector)
JugsawVector(vector::Vector)

Fields

head::Symbol

typename::String

fields::Vector

JugsawADT is an intermediate representation between Jugsaw IR data type and Julia native data type.

Examples

The Jugsaw object representation for 2+3im is

julia> JugsawObject("Base.ComplexF64", [2, 3])
JugsawADT(:Object, "Base.ComplexF64", [2, 3])

julia> JugsawVector([2, 3])
JugsawADT(:Vector, "", [2, 3])

JugsawIR.TypeTable — Type

struct TypeTable

The type definitions.

Fields

names::Vector{String}

defs::Dict{String, JDataType}

The defs defines a mapping from the type name to a JDataType instance.

Missing docstring.

Missing docstring for native2jugsaw. Check Documenter's build log for details.

Missing docstring.

Missing docstring for construct_object. Check Documenter's build log for details.