Jugsaw.Server

Jugsaw.Server.code_handler — Method

code_handler(
    req::HTTP.Messages.Request,
    app::AppSpecification
) -> HTTP.Messages.Response

Handle the request of generating the API for calling from a specific client language.

Response

[Success]: a JSON object with requested API code {"code" : ...}.

[NoDemoException]: a JSON object {"error" : ...}.

[ErrorException]: a JSON object {"error" : ...}.

Jugsaw.Server.demos_handler — Method

demos_handler(
    app::AppSpecification
) -> HTTP.Messages.Response

Handle the request of getting application specification, including registered function demos and type definitions.

Response

[Success]: Jugsaw IR in the form of a JSON object.

Jugsaw.Server.fetch_handler — Method

fetch_handler(
    r::Jugsaw.Server.AppRuntime,
    req::HTTP.Messages.Request
) -> HTTP.Messages.Response

Handle the request of fetching computed results and return a response with job id.

Request

A JSON payload that specifies the job id as {"job_id" : ...}.

Response

[Success]: Jugsaw IR in the form of JSON payload.

[TimedOutException]: a JSON object {"error" : ...}.

[ErrorException]: a JSON object {"error" : ...}.

Jugsaw.Server.fetch_status — Function

fetch_status(dapr::AbstractEventService, job_id::String; timeout::Real=get_timeout()) -> (status_code, status)

Get the status of a job. The return value is a tuple with the following two elements

status_code is a symbol to indicate the status query result, which can be :ok or :timed_out

status is a JobStatus object if the status_code is :ok, otherwise, is nothing.

Jugsaw.Server.get_query_interval — Method

get_query_interval() -> Any

Returns the query time interval of the event service in seconds.

Jugsaw.Server.get_timeout — Method

get_timeout() -> Any

Returns the network timeout of the event service access in seconds.

Jugsaw.Server.job_handler — Method

job_handler(
    r::Jugsaw.Server.AppRuntime,
    req::HTTP.Messages.Request
) -> HTTP.Messages.Response

Handle the request of function call and returns a response with job id.

Request

A Jugsaw IR that corresponds to a JobSpec instance.

Response

[Success]: a JSON object {"job_id" : ...}.

[NoDemoException]: a JSON object {"error" : ...}.

Jugsaw.Server.load_object — Method

load_object(dapr::AbstractEventService, job_id::AbstractString, resdemo; timeout::Real) -> (status_code, object)

Load an object to the main memory. The return value is a tuple with the following two elements

status_code is a symbol to indicate the status query result, which can be :ok or :timed_out

status is an object if the status_code is :ok, otherwise, is nothing.

The keyword argument timeout is should be greater than the expected job run time.

Jugsaw.Server.load_object_as_ir — Function

load_object_as_ir(dapr::AbstractEventService, job_id::AbstractString; timeout::Real) -> (status_code, ir)

Similar to load_object, but returns a Jugsaw IR instead. An object demo is not required.

Jugsaw.Server.open_in_default_browser — Method

open_in_default_browser(url)

Open a URL in the ambient default browser.

Note: this was copied from LiveServer.jl, and the original copy is from Pluto.jl.

Jugsaw.Server.publish_status — Function

publish_status(dapr::AbstractEventService, job_status::JobStatus) -> nothing

Publish the status of a job to the event service. The published event can be accessed with fetch_status function.

Jugsaw.Server.save_object — Function

save_object(dapr::AbstractEventService, job_id::AbstractString, res) -> nothing

Save an object to the event service in the form of local or web storage. The stored object can be loaded with load_object function.

Jugsaw.Server.serve — Method

serve(
    app::AppSpecification;
    eventservice,
    liveupdate,
    host,
    port,
    localurl,
    launch_browser,
    watched_files
) -> Union{Nothing, Task, HTTP.Servers.Server{HTTP.Servers.Listener{Nothing, Sockets.TCPServer}}}

Serve this application on specified host and port.

Arguments

app is a AppSpecification instance.

Keyword arguments

eventservice is a AbstractEventService instance, which is used to store job status and computed results.

The default value of eventservice depends on environment variable "JUGSAWSERVER". If the server runs on the localhost, then the value of "JUGSAWSERVER" should be "LOCAL" or missing, and default value of eventservice is InMemoryEventService(). Otherwise if the server runs on a docker container, then the value of "JUGSAW_SERVER" should be something else, and default value of eventservice is DaprService().

liveupdate is a boolean variable. If liveupdate is true, application will be updated automatically.

The default value of liveupdate depends on environment variable "JUGSAWSERVER". If the server runs on the localhost, then the value of "JUGSAWSERVER" should be "LOCAL" or missing, and default value of liveupdate is true. Otherwise if the server runs on a docker container, then the value of "JUGSAW_SERVER" should be something else, and default value of liveupdate is false.

watched_files is a list of file paths to watch with Revise.jl, the server will restart automatically on the change of watched files if liveserve is true.

host is the IP address or url of the host.

port is the port to serve the application.

launch_browser is boolean variable. If both this variable and liveserve are true, the default browser will open an html page for end-to-end testing.

localurl is a switch to serve in local mode with a simplified routing table.

In the local mode, the project name and application name are not required in the request url.

The route table

("GET", "/") -> get the index page (for local debugging).

("POST", "/v1/proj/{project}/app/{appname}/ver/{version}/func/{fname}") -> call a function and return a job id, please check job_handler.

("POST", "/v1/job/{jobid}/result") -> fetch results with a job id, please check [`fetchhandler`](@ref).

("GET", "/v1/proj/{project}/app/{appname}/ver/{version}/func") -> get application information, please check demos_handler.

("GET", "/v1/proj/{project}/app/{appname}/ver/{version}/func/{fname}/api/{lang}") -> get the API call for a client language, please check code_handler.

("GET", "/v1/proj/{project}/app/{appname}/ver/{version}/healthz") -> get the status of current application.

Jugsaw.Server.simpleserve — Method

simpleserve(runtime::AppRuntime; is_async=false, host="0.0.0.0", port=8088, localurl=false)

Serve this application on specified host and port.

Arguments

runtime is an AppRuntime instance.

Keyword arguments

is_async is a switch to turn on the asynchronous mode for debugging.

host is the IP address or url of the host.

port is the port to serve the application.

localurl is a switch to serve in local mode with a simplified routing table.

In the local mode, the project name and application name are not required in the request url.

Jugsaw.Server.AbstractEventService — Type

AbstractEventService

The abstract type for event service. Its concrete subtypes include

DaprService

FileEventService

InMemoryEventService

Required Interfaces

publish_status

fetch_status

save_object

load_object

load_object_as_ir

Jugsaw.Server.AppRuntime — Type

struct AppRuntime{ES<:Jugsaw.Server.AbstractEventService}

The application instance wrapped with run time information.

Fields

app is a AppSpecification instance.

dapr is a AbstractEventService instance for handling result storing and job status updating.

channel is a channel of jobs to be processed.

Jugsaw.Server.DaprService — Type

DaprService <: AbstractEventService

Dapr event service for storing and fetching events and results. Please check AbstractEventService for implemented interfaces.

Jugsaw.Server.FileEventService — Type

FileEventService <: AbstractEventService

Mocked event service for storing and fetching events and results from the local file system. Please check AbstractEventService for implemented interfaces.

Jugsaw.Server.InMemoryEventService — Type

struct InMemoryEventService <: Jugsaw.Server.AbstractEventService

An event service for storing and fetching events and results from the the main memory. Please check AbstractEventService for implemented interfaces.

Fields

print_event::Bool

object_store::Dict{String, Any}

status_store::Dict{String, Jugsaw.Server.JobStatus}

Jugsaw.Server.Job — Type

struct Job

A resolved job can be queued and executed in a AppRuntime.

Fields

id::String

created_at::Float64

created_by::String

maxtime::Float64

demo::JugsawDemo

args::Tuple

kwargs::NamedTuple

Jugsaw.Server.JobSpec — Type

struct JobSpec

A job with function payload specified as a JugsawExpr.

Fields

id::String

created_at::Float64

created_by::String

maxtime::Float64

fname::String

args::JugsawExpr

kwargs::JugsawExpr

Here id is the job id that used to store and fetch computed results.

Jugsaw.Server.JobStatus — Type

struct JobStatus

A job status that can be pubished to AbstractEventService.

Fields

id::String

status::Jugsaw.Server.JobStatusEnum

timestamp::Float64

description::String