Jugsaw

Chained function call

When Jugsaw server gets a chained function call, like sin(cos(0.5)). The following two tasks will be added to the task queue.

Call(cos, (0.5,), (;)) -> id1
Call(sin, (object_getter(state_store, id1),), (;))

where -> points to the id of the returned object in the state_store. The state_store is a dictionary mapping an object id to its value. When querying an object from the state_store, the program waits for the corresponding task to complete.

object_getter(id) returns a Call instance with the following definition

function object_getter(state_store::StateStore, object_id::String)
    Call((s, id)->Meta.parse(Base.getindex(s, id)), (state_store, object_id), (;))
end

The nested Call is then executed by the JugsawIR.fevalself with the following steps

sin function is triggered,

while rendering the arguments of sin, the object getter(Call) will trigger the state_store[id1],

wait for the cos function to complete,

with the returned object, execute the sin function.

APIs

Jugsaw.generate_code — Method

generate_code(lang, endpoint::String, appname::Symbol, fcall::JugsawADT, democall::JugsawIR.Call)

Generate code for target language.

Arguments

lang can be a string or an AbstractLang instance that specifies the target language.

Please use subtypes(AbstractLang) for supported client languages.

endpoint is the url for service provider, e.g. it can be https://www.jugsaw.co.

appname is the application name.

fcall is a JugsawADT that specifies the function call.

democall is the demo instance of that function call.